+++ src/main/java/com/asg/shipping/importmanifestbl/service/ImportManifestService.java
@@ -7,6 +7,7 @@ import com.asg.shipping.importmanifestupdate.dto.*;
 
 import org.springframework.data.domain.Pageable;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Map;
 
@@ -22,7 +23,7 @@ public interface ImportManifestService {
 
     SendEdiEmailsResponseDto sendEdiEmails(Long transactionPoId);
 
-    LoadEmailFaxResponseDto loadEmailFax(Long addressMasterPoid, String addressType);
+    LoadEmailFaxResponseDto loadEmailFax(BigDecimal addressMasterPoid, String addressType);
 
     BlStatusResponseDto getBlStatus(Long transactionPoId);
